Monster nav: multi-layer pathfinding & debug overhaul
[General]: Improved nav layer struct to use 32-bit flags and clearance, enabling robust sampling for tall ceilings and complex environments.
[General]: Enhanced step traversal validation to segment long climbs and ramps, mirroring PMove constraints for reliable stair/ramp navigation.
[General]: Extensive diagnostics: added debug output for step-test failures, edge rejections (drop cap, clearance, slope), and A* stagnation/time budget.
[General]: Navigation debug draw now visualizes rejected edges for clearance, drop cap, and slope; activated at end of server frame for up-to-date rendering.
[General]: Tuned path policy parameters (step height, drop cap, waypoint radius, rebuild intervals) for more responsive movement.

[SVGame]: A* pathfinding now expands neighbors vertically (Z), including diagonal and multi-cell ramp transitions, allowing monsters to traverse stacked layers and stairs.
[SVGame]: Monsters now prefer breadcrumb/trail goals when target is out of LOS or on another floor, avoiding repeated pathing failures to unreachable points.
[SVGame]: Path rebuilds are throttled and backed off more aggressively; logic improved to keep following existing paths when rebuilds are skipped.
[SVGame]: Improved handling of edge cases in pathfinding and step validation; minor bug fixes and comment improvements.

[CLGame]: Debug entity event printing now gated by cvar; default debug entity events cvar set to 0.

Overall, these changes make monster navigation much more reliable in multi-layered environments, provide better diagnostics for navmesh issues, and ensure monsters can traverse stairs, ramps, and overlapping floors without getting stuck.

+++ src/baseq2rtxp/svgame/nav/svg_nav.h
@@ -54,9 +54,9 @@ typedef struct nav_layer_s {
     //! Quantized Z position.
     int16_t z_quantized;
     //! Content flags (nav_layer_flags_t).
-    uint8_t flags;
-    //! Optional clearance in grid cells.
-    uint8_t clearance;
+    uint32_t flags;
+	//! Optional clearance in grid cells (quantized, stored with wider range for tall ceilings).
+	uint32_t clearance;
 } nav_layer_t;
 
 /**
